renamed project and changed folder structure
