--- docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-m1t1.zip/com.acme.m1t1.web.internal.portlet.M1T1Portlet.config
@@ -1,1 +1,0 @@
-_m1T1.target="(component.name\=com.acme.m1t1.internal.custom.M1T1CustomImpl)"
