--- docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-m1t1.zip/m1t1-api/src/main/java/com/acme/m1t1/M1T1.java
@@ -1,7 +1,0 @@
-package com.acme.m1t1;
-
-public interface M1T1 {
-
-	public String doSomething();
-
-}
