--- docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-m1t1.zip/m1t1-custom-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Acme M1T1 Custom Implementation
-Bundle-SymbolicName: com.acme.m1t1.custom.impl
-Bundle-Version: 1.0.0
