--- docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-m1t1.zip/m1t1-custom-impl/src/main/java/com/acme/m1t1/internal/custom/M1T1CustomImpl.java
@@ -1,28 +1,0 @@
-package com.acme.m1t1.internal.custom;
-
-import com.acme.m1t1.M1T1;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-@Component(service = M1T1.class)
-public class M1T1CustomImpl implements M1T1 {
-
-	@Override
-	public String doSomething() {
-		StringBuilder sb = new StringBuilder();
-
-		sb.append(
-			"M1T1CustomImpl, which delegates to " +
-				_defaultService.doSomething());
-
-		return sb.toString();
-	}
-
-	@Reference(
-		target = "(component.name=com.acme.m1t1.internal.M1T1Impl)",
-		unbind = "-"
-	)
-	private M1T1 _defaultService;
-
-}
