+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-t4n9.zip/com.acme.t4n9.web.internal.portlet.T4N9Portlet.config
@@ -1,0 +1,1 @@
+_t4N9.target="(component.name\=com.acme.t4n9.baker.internal.T4N9BakerImpl)"
