+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-t4n9.zip/t4n9-able-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Acme T4N9 Able Implementation
+Bundle-SymbolicName: com.acme.t4n9.able.impl
+Bundle-Version: 1.0.0
