+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-t4n9.zip/t4n9-able-impl/src/main/java/com/acme/t4n9/able/internal/T4N9AbleImpl.java
@@ -1,0 +1,15 @@
+package com.acme.t4n9.able.internal;
+
+import com.acme.t4n9.T4N9;
+
+import org.osgi.service.component.annotations.Component;
+
+@Component(service = T4N9.class)
+public class T4N9AbleImpl implements T4N9 {
+
+	@Override
+	public String doSomething() {
+		return "This is the T4N9 Able implementation.";
+	}
+
+}
