+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-t4n9.zip/t4n9-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Acme T4N9 API
+Bundle-SymbolicName: com.acme.t4n9.api
+Bundle-Version: 1.0.0
+Export-Package: com.acme.t4n9
