+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-t4n9.zip/t4n9-api/src/main/java/com/acme/t4n9/T4N9.java
@@ -1,0 +1,10 @@
+package com.acme.t4n9;
+
+import aQute.bnd.annotation.ProviderType;
+
+@ProviderType
+public interface T4N9 {
+
+	public String doSomething();
+
+}
