+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-t4n9.zip/t4n9-baker-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Acme M1T1 Baker Implementation
+Bundle-SymbolicName: com.acme.t4n9.baker.impl
+Bundle-Version: 1.0.0
