+++ docs/dxp/7.x/en/liferay-internals/extending-liferay/overriding-osgi-services/reconfiguring-components-to-use-your-osgi-service/resources/liferay-t4n9.zip/t4n9-baker-impl/src/main/java/com/acme/t4n9/baker/internal/T4N9BakerImpl.java
@@ -1,0 +1,23 @@
+package com.acme.t4n9.baker.internal;
+
+import com.acme.t4n9.T4N9;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+@Component(property = "service.ranking:Integer=100", service = T4N9.class)
+public class T4N9BakerImpl implements T4N9 {
+
+	@Override
+	public String doSomething() {
+		return _t4n9.doSomething() +
+			"<br />This is the T4N9 Baker implementation.";
+	}
+
+	@Reference(
+		target = "(component.name=com.acme.t4n9.able.internal.T4N9AbleImpl)",
+		unbind = "-"
+	)
+	private T4N9 _t4n9;
+
+}
